1. 设置新增终端、编辑器字体设置

+++ fx-plus/src/main/java/cn/oyzh/fx/plus/font/FontWeightComboBox.java
@@ -38,7 +38,7 @@ public class FontWeightComboBox extends FlexComboBox<FontWeight> {
         }
     }
 
-    public void selectWeight(Short fontWeight) {
+    public void selectWeight(Integer fontWeight) {
         if (fontWeight == null) {
             this.select(null);
         } else {
